server/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.7.1=pyhd8ed1ab_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.5.7=hf0a4a13_0
click=8.0.4=py311hca03da5_0
exceptiongroup=1.1.2=pyhd8ed1ab_0
fastapi=0.100.0=pyhd8ed1ab_0
h11=0.12.0=pyhd3eb1b0_0
idna=3.4=pyhd8ed1ab_0
libffi=3.4.4=hca03da5_0
ncurses=6.4=h313beb8_0
openssl=3.1.1=h53f4e23_1
paho-mqtt=1.6.1=pyhd8ed1ab_0
pip=23.1.2=py311hca03da5_0
pydantic=1.10.8=py311h80987f9_0
python=3.11.4=hb885b13_0
readline=8.2=h1a28f6b_0
setuptools=67.8.0=py311hca03da5_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.41.2=h80987f9_0
starlette=0.27.0=pyhd8ed1ab_0
tk=8.6.12=hb8d0fd4_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
uvicorn=0.20.0=py311hca03da5_0
websockets=10.4=py311h80987f9_1
wheel=0.38.4=py311hca03da5_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
